requirements.txt
Flask
numpy
pandas
pdfminer.six
python-docx
spacy